+++ example/comparison2.py
@@ -43,14 +43,15 @@ for it in range(n_matrix_pairs):
     cols = rng1.randint(nr_vocab, size=nnz_a)
     data = rng1.rand(nnz_a)
     
-    a_sparse = coo_matrix((data, (row, cols)), shape=(n_samples, nr_vocab))
+    dtype = np.float32
+    a_sparse = coo_matrix((data, (row, cols)), shape=(n_samples, nr_vocab), dtype=dtype)
     a = a_sparse.tocsr()
     
     row = rng1.randint(n_duplicates, size=nnz_b)
     cols = rng1.randint(nr_vocab, size=nnz_b)
     data = rng1.rand(nnz_b)
     
-    b_sparse = coo_matrix((data, (row, cols)), shape=(n_duplicates, nr_vocab))
+    b_sparse = coo_matrix((data, (row, cols)), shape=(n_duplicates, nr_vocab), dtype=dtype)
     b = b_sparse.T.tocsr()
     
     C, C_ntop = awesome_cossim_topn(a, b, N, thresh, return_best_ntop=True)